Module google_dialogflow3::api

Structs

Dialogflow
Central instance to access all Dialogflow related resource activities

GoogleCloudDialogflowCxV3Agent
Agents are best described as Natural Language Understanding (NLU) modules that transform user requests into actionable data. You can include agents in your app, product, or service to determine user intent and respond to the user in a natural way. After you create an agent, you can add Intents, Entity Types, Flows, Fulfillments, Webhooks, and so on to manage the conversation flows..

GoogleCloudDialogflowCxV3AgentValidationResult
The response message for Agents.GetAgentValidationResult.

GoogleCloudDialogflowCxV3AudioInput
Represents the natural speech audio to be processed.

GoogleCloudDialogflowCxV3BatchDeleteTestCasesRequest
The request message for TestCases.BatchDeleteTestCases.

GoogleCloudDialogflowCxV3BatchRunTestCasesRequest
The request message for TestCases.BatchRunTestCases.

GoogleCloudDialogflowCxV3CalculateCoverageResponse
The response message for TestCases.CalculateCoverage.

GoogleCloudDialogflowCxV3ConversationTurn
One interaction between a human and virtual agent. The human provides some input and the virtual agent provides a response.

GoogleCloudDialogflowCxV3ConversationTurnUserInput
The input from the human user.

GoogleCloudDialogflowCxV3ConversationTurnVirtualAgentOutput
The output from the virtual agent.

GoogleCloudDialogflowCxV3DetectIntentRequest
The request to detect user's intent.

GoogleCloudDialogflowCxV3DetectIntentResponse
The message returned from the DetectIntent method.

GoogleCloudDialogflowCxV3DtmfInput
Represents the input for dtmf event.

GoogleCloudDialogflowCxV3EntityType
Entities are extracted from user input and represent parameters that are meaningful to your application. For example, a date range, a proper name such as a geographic location or landmark, and so on. Entities represent actionable data for your application. When you define an entity, you can also include synonyms that all map to that entity. For example, "soft drink", "soda", "pop", and so on. There are three types of entities: * System - entities that are defined by the Dialogflow API for common data types such as date, time, currency, and so on. A system entity is represented by the EntityType type. * Custom - entities that are defined by you that represent actionable data that is meaningful to your application. For example, you could define a pizza.sauce entity for red or white pizza sauce, a pizza.cheese entity for the different types of cheese on a pizza, a pizza.topping entity for different toppings, and so on. A custom entity is represented by the EntityType type. * User - entities that are built for an individual user such as favorites, preferences, playlists, and so on. A user entity is represented by the SessionEntityType type. For more information about entity types, see the Dialogflow documentation.

GoogleCloudDialogflowCxV3EntityTypeEntity
An entity entry for an associated entity type.

GoogleCloudDialogflowCxV3EntityTypeExcludedPhrase
An excluded entity phrase that should not be matched.

GoogleCloudDialogflowCxV3Environment
Represents an environment for an agent. You can create multiple versions of your agent and publish them to separate environments. When you edit an agent, you are editing the draft agent. At any point, you can save the draft agent as an agent version, which is an immutable snapshot of your agent. When you save the draft agent, it is published to the default environment. When you create agent versions, you can publish them to custom environments. You can create a variety of custom environments for testing, development, production, etc.

GoogleCloudDialogflowCxV3EnvironmentVersionConfig
Configuration for the version.

GoogleCloudDialogflowCxV3EventHandler
An event handler specifies an event that can be handled during a session. When the specified event happens, the following actions are taken in order: * If there is a trigger_fulfillment associated with the event, it will be called. * If there is a target_page associated with the event, the session will transition into the specified page. * If there is a target_flow associated with the event, the session will transition into the specified flow.

GoogleCloudDialogflowCxV3EventInput
Represents the event to trigger.

GoogleCloudDialogflowCxV3Experiment
Represents an experiment in an environment.

GoogleCloudDialogflowCxV3ExperimentDefinition
Definition of the experiment.

GoogleCloudDialogflowCxV3ExperimentResult
The inference result which includes an objective metric to optimize and the confidence interval.

GoogleCloudDialogflowCxV3ExperimentResultConfidenceInterval
A confidence interval is a range of possible values for the experiment objective you are trying to measure.

GoogleCloudDialogflowCxV3ExperimentResultMetric
Metric and corresponding confidence intervals.

GoogleCloudDialogflowCxV3ExperimentResultVersionMetrics
Version variant and associated metrics.

GoogleCloudDialogflowCxV3ExportAgentRequest
The request message for Agents.ExportAgent.

GoogleCloudDialogflowCxV3ExportTestCasesRequest
The request message for TestCases.ExportTestCases.

GoogleCloudDialogflowCxV3Flow
Flows represents the conversation flows when you build your chatbot agent. A flow consists of many pages connected by the transition routes. Conversations always start with the built-in Start Flow (with an all-0 ID). Transition routes can direct the conversation session from the current flow (parent flow) to another flow (sub flow). When the sub flow is finished, Dialogflow will bring the session back to the parent flow, where the sub flow is started. Usually, when a transition route is followed by a matched intent, the intent will be "consumed". This means the intent won't activate more transition routes. However, when the followed transition route moves the conversation session into a different flow, the matched intent can be carried over and to be consumed in the target flow.

GoogleCloudDialogflowCxV3FlowValidationResult
The response message for Flows.GetFlowValidationResult.

GoogleCloudDialogflowCxV3Form
A form is a data model that groups related parameters that can be collected from the user. The process in which the agent prompts the user and collects parameter values from the user is called form filling. A form can be added to a page. When form filling is done, the filled parameters will be written to the session.

GoogleCloudDialogflowCxV3FormParameter
Represents a form parameter.

GoogleCloudDialogflowCxV3FormParameterFillBehavior
Configuration for how the filling of a parameter should be handled.

GoogleCloudDialogflowCxV3FulfillIntentRequest
Request of FulfillIntent

GoogleCloudDialogflowCxV3FulfillIntentResponse
Response of FulfillIntent

GoogleCloudDialogflowCxV3Fulfillment
A fulfillment can do one or more of the following actions at the same time: * Generate rich message responses. * Set parameter values. * Call the webhook. Fulfillments can be called at various stages in the Page or Form lifecycle. For example, when a DetectIntentRequest drives a session to enter a new page, the page's entry fulfillment can add a static response to the QueryResult in the returning DetectIntentResponse, call the webhook (for example, to load user data from a database), or both.

GoogleCloudDialogflowCxV3FulfillmentConditionalCases
A list of cascading if-else conditions. Cases are mutually exclusive. The first one with a matching condition is selected, all the rest ignored.

GoogleCloudDialogflowCxV3FulfillmentConditionalCasesCase
Each case has a Boolean condition. When it is evaluated to be True, the corresponding messages will be selected and evaluated recursively.

GoogleCloudDialogflowCxV3FulfillmentConditionalCasesCaseCaseContent
The list of messages or conditional cases to activate for this case.

GoogleCloudDialogflowCxV3FulfillmentSetParameterAction
Setting a parameter value.

GoogleCloudDialogflowCxV3ImportTestCasesRequest
The request message for TestCases.ImportTestCases.

GoogleCloudDialogflowCxV3InputAudioConfig
Instructs the speech recognizer on how to process the audio content.

GoogleCloudDialogflowCxV3Intent
An intent represents a user's intent to interact with a conversational agent. You can provide information for the Dialogflow API to use to match user input to an intent by adding training phrases (i.e., examples of user input) to your intent.

GoogleCloudDialogflowCxV3IntentCoverage
Intent coverage represents the percentage of all possible intents in the agent that are triggered in any of a parent's test cases.

GoogleCloudDialogflowCxV3IntentCoverageIntent
The agent's intent.

GoogleCloudDialogflowCxV3IntentInput
Represents the intent to trigger programmatically rather than as a result of natural language processing.

GoogleCloudDialogflowCxV3IntentParameter
Represents an intent parameter.

GoogleCloudDialogflowCxV3IntentTrainingPhrase
Represents an example that the agent is trained on to identify the intent.

GoogleCloudDialogflowCxV3IntentTrainingPhrasePart
Represents a part of a training phrase.

GoogleCloudDialogflowCxV3ListAgentsResponse
The response message for Agents.ListAgents.

GoogleCloudDialogflowCxV3ListEntityTypesResponse
The response message for EntityTypes.ListEntityTypes.

GoogleCloudDialogflowCxV3ListEnvironmentsResponse
The response message for Environments.ListEnvironments.

GoogleCloudDialogflowCxV3ListExperimentsResponse
The response message for Experiments.ListExperiments.

GoogleCloudDialogflowCxV3ListFlowsResponse
The response message for Flows.ListFlows.

GoogleCloudDialogflowCxV3ListIntentsResponse
The response message for Intents.ListIntents.

GoogleCloudDialogflowCxV3ListPagesResponse
The response message for Pages.ListPages.

GoogleCloudDialogflowCxV3ListSecuritySettingsResponse
The response message for SecuritySettings.ListSecuritySettings.

GoogleCloudDialogflowCxV3ListSessionEntityTypesResponse
The response message for SessionEntityTypes.ListSessionEntityTypes.

GoogleCloudDialogflowCxV3ListTestCaseResultsResponse
The response message for TestCases.ListTestCaseResults.

GoogleCloudDialogflowCxV3ListTestCasesResponse
The response message for TestCases.ListTestCases.

GoogleCloudDialogflowCxV3ListTransitionRouteGroupsResponse
The response message for TransitionRouteGroups.ListTransitionRouteGroups.

GoogleCloudDialogflowCxV3ListVersionsResponse
The response message for Versions.ListVersions.

GoogleCloudDialogflowCxV3ListWebhooksResponse
The response message for Webhooks.ListWebhooks.

GoogleCloudDialogflowCxV3LoadVersionRequest
The request message for Versions.LoadVersion.

GoogleCloudDialogflowCxV3LookupEnvironmentHistoryResponse
The response message for Environments.LookupEnvironmentHistory.

GoogleCloudDialogflowCxV3Match
Represents one match result of MatchIntent.

GoogleCloudDialogflowCxV3MatchIntentRequest
Request of MatchIntent.

GoogleCloudDialogflowCxV3MatchIntentResponse
Response of MatchIntent.

GoogleCloudDialogflowCxV3NluSettings
Settings related to NLU.

GoogleCloudDialogflowCxV3OutputAudioConfig
Instructs the speech synthesizer how to generate the output audio content.

GoogleCloudDialogflowCxV3Page
A Dialogflow CX conversation (session) can be described and visualized as a state machine. The states of a CX session are represented by pages. For each flow, you define many pages, where your combined pages can handle a complete conversation on the topics the flow is designed for. At any given moment, exactly one page is the current page, the current page is considered active, and the flow associated with that page is considered active. Every flow has a special start page. When a flow initially becomes active, the start page page becomes the current page. For each conversational turn, the current page will either stay the same or transition to another page. You configure each page to collect information from the end-user that is relevant for the conversational state represented by the page. For more information, see the Page guide.

GoogleCloudDialogflowCxV3QueryInput
Represents the query input. It can contain one of: 1. A conversational query in the form of text. 2. An intent query that specifies which intent to trigger. 3. Natural language speech audio to be processed. 4. An event to be triggered.

GoogleCloudDialogflowCxV3QueryParameters
Represents the parameters of a conversational query.

GoogleCloudDialogflowCxV3QueryResult
Represents the result of a conversational query.

GoogleCloudDialogflowCxV3ResourceName
Resource name and display name.

GoogleCloudDialogflowCxV3ResponseMessage
Represents a response message that can be returned by a conversational agent. Response messages are also used for output audio synthesis. The approach is as follows: * If at least one OutputAudioText response is present, then all OutputAudioText responses are linearly concatenated, and the result is used for output audio synthesis. * If the OutputAudioText responses are a mixture of text and SSML, then the concatenated result is treated as SSML; otherwise, the result is treated as either text or SSML as appropriate. The agent designer should ideally use either text or SSML consistently throughout the bot design. * Otherwise, all Text responses are linearly concatenated, and the result is used for output audio synthesis. This approach allows for more sophisticated user experience scenarios, where the text displayed to the user may differ from what is heard.

GoogleCloudDialogflowCxV3ResponseMessageConversationSuccess
Indicates that the conversation succeeded, i.e., the bot handled the issue that the customer talked to it about. Dialogflow only uses this to determine which conversations should be counted as successful and doesn't process the metadata in this message in any way. Note that Dialogflow also considers conversations that get to the conversation end page as successful even if they don't return ConversationSuccess. You may set this, for example: * In the entry_fulfillment of a Page if entering the page indicates that the conversation succeeded. * In a webhook response when you determine that you handled the customer issue.

GoogleCloudDialogflowCxV3ResponseMessageEndInteraction
Indicates that interaction with the Dialogflow agent has ended. This message is generated by Dialogflow only and not supposed to be defined by the user.

GoogleCloudDialogflowCxV3ResponseMessageLiveAgentHandoff
Indicates that the conversation should be handed off to a live agent. Dialogflow only uses this to determine which conversations were handed off to a human agent for measurement purposes. What else to do with this signal is up to you and your handoff procedures. You may set this, for example: * In the entry_fulfillment of a Page if entering the page indicates something went extremely wrong in the conversation. * In a webhook response when you determine that the customer issue can only be handled by a human.

GoogleCloudDialogflowCxV3ResponseMessageMixedAudio
Represents an audio message that is composed of both segments synthesized from the Dialogflow agent prompts and ones hosted externally at the specified URIs. The external URIs are specified via play_audio. This message is generated by Dialogflow only and not supposed to be defined by the user.

GoogleCloudDialogflowCxV3ResponseMessageMixedAudioSegment
Represents one segment of audio.

GoogleCloudDialogflowCxV3ResponseMessageOutputAudioText
A text or ssml response that is preferentially used for TTS output audio synthesis, as described in the comment on the ResponseMessage message.

GoogleCloudDialogflowCxV3ResponseMessagePlayAudio
Specifies an audio clip to be played by the client as part of the response.

GoogleCloudDialogflowCxV3ResponseMessageText
The text response message.

GoogleCloudDialogflowCxV3RestoreAgentRequest
The request message for Agents.RestoreAgent.

GoogleCloudDialogflowCxV3RunTestCaseRequest
The request message for TestCases.RunTestCase.

GoogleCloudDialogflowCxV3SecuritySettings
Represents the settings related to security issues, such as data redaction and data retention. It may take hours for updates on the settings to propagate to all the related components and take effect.

GoogleCloudDialogflowCxV3SentimentAnalysisResult
The result of sentiment analysis. Sentiment analysis inspects user input and identifies the prevailing subjective opinion, especially to determine a user's attitude as positive, negative, or neutral.

GoogleCloudDialogflowCxV3SessionEntityType
Session entity types are referred to as User entity types and are entities that are built for an individual user such as favorites, preferences, playlists, and so on. You can redefine a session entity type at the session level to extend or replace a custom entity type at the user session level (we refer to the entity types defined at the agent level as "custom entity types"). Note: session entity types apply to all queries, regardless of the language. For more information about entity types, see the Dialogflow documentation.

GoogleCloudDialogflowCxV3SpeechToTextSettings
Settings related to speech recognition.

GoogleCloudDialogflowCxV3StartExperimentRequest
The request message for Experiments.StartExperiment.

GoogleCloudDialogflowCxV3StopExperimentRequest
The request message for Experiments.StopExperiment.

GoogleCloudDialogflowCxV3SynthesizeSpeechConfig
Configuration of how speech should be synthesized.

GoogleCloudDialogflowCxV3TestCase
Represents a test case.

GoogleCloudDialogflowCxV3TestCaseResult
Represents a result from running a test case in an agent environment.

GoogleCloudDialogflowCxV3TestConfig
Represents configurations for a test case.

GoogleCloudDialogflowCxV3TestRunDifference
The description of differences between original and replayed agent output.

GoogleCloudDialogflowCxV3TextInput
Represents the natural language text to be processed.

GoogleCloudDialogflowCxV3TrainFlowRequest
The request message for Flows.TrainFlow.

GoogleCloudDialogflowCxV3TransitionCoverage
Transition coverage represents the percentage of all possible page transitions (page-level transition routes and event handlers, excluding transition route groups) present within any of a parent's test cases.

GoogleCloudDialogflowCxV3TransitionCoverageTransition
A transition in a page.

GoogleCloudDialogflowCxV3TransitionCoverageTransitionNode
The source or target of a transition.

GoogleCloudDialogflowCxV3TransitionRoute
A transition route specifies a intent that can be matched and/or a data condition that can be evaluated during a session. When a specified transition is matched, the following actions are taken in order: * If there is a trigger_fulfillment associated with the transition, it will be called. * If there is a target_page associated with the transition, the session will transition into the specified page. * If there is a target_flow associated with the transition, the session will transition into the specified flow.

GoogleCloudDialogflowCxV3TransitionRouteGroup
An TransitionRouteGroup represents a group of TransitionRoutes to be used by a Page.

GoogleCloudDialogflowCxV3TransitionRouteGroupCoverage
Transition route group coverage represents the percentage of all possible transition routes present within any of a parent's test cases. The results are grouped by the transition route group.

GoogleCloudDialogflowCxV3TransitionRouteGroupCoverageCoverage
Coverage result message for one transition route group.

GoogleCloudDialogflowCxV3TransitionRouteGroupCoverageCoverageTransition
A transition coverage in a transition route group.

GoogleCloudDialogflowCxV3ValidateAgentRequest
The request message for Agents.ValidateAgent.

GoogleCloudDialogflowCxV3ValidateFlowRequest
The request message for Flows.ValidateFlow.

GoogleCloudDialogflowCxV3ValidationMessage
Agent/flow validation message.

GoogleCloudDialogflowCxV3VariantsHistory
The history of variants update.

GoogleCloudDialogflowCxV3Version
Represents a version of a flow.

GoogleCloudDialogflowCxV3VersionVariants
A list of flow version variants.

GoogleCloudDialogflowCxV3VersionVariantsVariant
A single flow version with specified traffic allocation.

GoogleCloudDialogflowCxV3VoiceSelectionParams
Description of which voice to use for speech synthesis.

GoogleCloudDialogflowCxV3Webhook
Webhooks host the developer's business logic. During a session, webhooks allow the developer to use the data extracted by Dialogflow's natural language processing to generate dynamic responses, validate collected data, or trigger actions on the backend.

GoogleCloudDialogflowCxV3WebhookGenericWebService
Represents configuration for a generic web service.

GoogleLongrunningListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

GoogleTypeLatLng
An object that represents a latitude/longitude pair. This is expressed as a pair of doubles to represent degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

ProjectLocationAgentCreateCall
Creates an agent in the specified location.

ProjectLocationAgentDeleteCall
Deletes the specified agent.

ProjectLocationAgentEntityTypeCreateCall
Creates an entity type in the specified agent.

ProjectLocationAgentEntityTypeDeleteCall
Deletes the specified entity type.

ProjectLocationAgentEntityTypeGetCall
Retrieves the specified entity type.

ProjectLocationAgentEntityTypeListCall
Returns the list of all entity types in the specified agent.

ProjectLocationAgentEntityTypePatchCall
Updates the specified entity type.

ProjectLocationAgentEnvironmentCreateCall
Creates an Environment in the specified Agent.

ProjectLocationAgentEnvironmentDeleteCall
Deletes the specified Environment.

ProjectLocationAgentEnvironmentExperimentCreateCall
Creates an Experiment in the specified Environment.

ProjectLocationAgentEnvironmentExperimentDeleteCall
Deletes the specified Experiment.

ProjectLocationAgentEnvironmentExperimentGetCall
Retrieves the specified Experiment.

ProjectLocationAgentEnvironmentExperimentListCall
Returns the list of all experiments in the specified Environment.

ProjectLocationAgentEnvironmentExperimentPatchCall
Updates the specified Experiment.

ProjectLocationAgentEnvironmentExperimentStartCall
Starts the specified Experiment. This rpc only changes the state of experiment from PENDING to RUNNING.

ProjectLocationAgentEnvironmentExperimentStopCall
Stops the specified Experiment. This rpc only changes the state of experiment from RUNNING to DONE.

ProjectLocationAgentEnvironmentGetCall
Retrieves the specified Environment.

ProjectLocationAgentEnvironmentListCall
Returns the list of all environments in the specified Agent.

ProjectLocationAgentEnvironmentLookupEnvironmentHistoryCall
Looks up the history of the specified Environment.

ProjectLocationAgentEnvironmentPatchCall
Updates the specified Environment.

ProjectLocationAgentEnvironmentSessionDetectIntentCall
Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause session entity types to be updated, which in turn might affect results of future queries. Note: Always use agent versions for production traffic. See Versions and environments.

ProjectLocationAgentEnvironmentSessionEntityTypeCreateCall
Creates a session entity type.

ProjectLocationAgentEnvironmentSessionEntityTypeDeleteCall
Deletes the specified session entity type.

ProjectLocationAgentEnvironmentSessionEntityTypeGetCall
Retrieves the specified session entity type.

ProjectLocationAgentEnvironmentSessionEntityTypeListCall
Returns the list of all session entity types in the specified session.

ProjectLocationAgentEnvironmentSessionEntityTypePatchCall
Updates the specified session entity type.

ProjectLocationAgentEnvironmentSessionFulfillIntentCall
Fulfills a matched intent returned by MatchIntent. Must be called after MatchIntent, with input from MatchIntentResponse. Otherwise, the behavior is undefined.

ProjectLocationAgentEnvironmentSessionMatchIntentCall
Returns preliminary intent match results, doesn't change the session status.

ProjectLocationAgentExportCall
Exports the specified agent to a binary file.

ProjectLocationAgentFlowCreateCall
Creates a flow in the specified agent.

ProjectLocationAgentFlowDeleteCall
Deletes a specified flow.

ProjectLocationAgentFlowGetCall
Retrieves the specified flow.

ProjectLocationAgentFlowGetValidationResultCall
Gets the latest flow validation result. Flow validation is performed when ValidateFlow is called.

ProjectLocationAgentFlowListCall
Returns the list of all flows in the specified agent.

ProjectLocationAgentFlowPageCreateCall
Creates a page in the specified flow.

ProjectLocationAgentFlowPageDeleteCall
Deletes the specified page.

ProjectLocationAgentFlowPageGetCall
Retrieves the specified page.

ProjectLocationAgentFlowPageListCall
Returns the list of all pages in the specified flow.

ProjectLocationAgentFlowPagePatchCall
Updates the specified page.

ProjectLocationAgentFlowPatchCall
Updates the specified flow.

ProjectLocationAgentFlowTrainCall
Trains the specified flow. Note that only the flow in 'draft' environment is trained.

ProjectLocationAgentFlowTransitionRouteGroupCreateCall
Creates an TransitionRouteGroup in the specified flow.

ProjectLocationAgentFlowTransitionRouteGroupDeleteCall
Deletes the specified TransitionRouteGroup.

ProjectLocationAgentFlowTransitionRouteGroupGetCall
Retrieves the specified TransitionRouteGroup.

ProjectLocationAgentFlowTransitionRouteGroupListCall
Returns the list of all transition route groups in the specified flow.

ProjectLocationAgentFlowTransitionRouteGroupPatchCall
Updates the specified TransitionRouteGroup.

ProjectLocationAgentFlowValidateCall
Validates the specified flow and creates or updates validation results. Please call this API after the training is completed to get the complete validation results.

ProjectLocationAgentFlowVersionCreateCall
Creates a Version in the specified Flow.

ProjectLocationAgentFlowVersionDeleteCall
Deletes the specified Version.

ProjectLocationAgentFlowVersionGetCall
Retrieves the specified Version.

ProjectLocationAgentFlowVersionListCall
Returns the list of all versions in the specified Flow.

ProjectLocationAgentFlowVersionLoadCall
Loads a specified version to draft version.

ProjectLocationAgentFlowVersionPatchCall
Updates the specified Version.

ProjectLocationAgentGetCall
Retrieves the specified agent.

ProjectLocationAgentGetValidationResultCall
Gets the latest agent validation result. Agent validation is performed when ValidateAgent is called.

ProjectLocationAgentIntentCreateCall
Creates an intent in the specified agent.

ProjectLocationAgentIntentDeleteCall
Deletes the specified intent.

ProjectLocationAgentIntentGetCall
Retrieves the specified intent.

ProjectLocationAgentIntentListCall
Returns the list of all intents in the specified agent.

ProjectLocationAgentIntentPatchCall
Updates the specified intent.

ProjectLocationAgentListCall
Returns the list of all agents in the specified location.

ProjectLocationAgentPatchCall
Updates the specified agent.

ProjectLocationAgentRestoreCall
Restores the specified agent from a binary file. Replaces the current agent with a new one. Note that all existing resources in agent (e.g. intents, entity types, flows) will be removed.

ProjectLocationAgentSessionDetectIntentCall
Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause session entity types to be updated, which in turn might affect results of future queries. Note: Always use agent versions for production traffic. See Versions and environments.

ProjectLocationAgentSessionEntityTypeCreateCall
Creates a session entity type.

ProjectLocationAgentSessionEntityTypeDeleteCall
Deletes the specified session entity type.

ProjectLocationAgentSessionEntityTypeGetCall
Retrieves the specified session entity type.

ProjectLocationAgentSessionEntityTypeListCall
Returns the list of all session entity types in the specified session.

ProjectLocationAgentSessionEntityTypePatchCall
Updates the specified session entity type.

ProjectLocationAgentSessionFulfillIntentCall
Fulfills a matched intent returned by MatchIntent. Must be called after MatchIntent, with input from MatchIntentResponse. Otherwise, the behavior is undefined.

ProjectLocationAgentSessionMatchIntentCall
Returns preliminary intent match results, doesn't change the session status.

ProjectLocationAgentTestCaseBatchDeleteCall
Batch deletes test cases.

ProjectLocationAgentTestCaseBatchRunCall
Kicks off a batch run of test cases.

ProjectLocationAgentTestCaseCalculateCoverageCall
Calculates the test coverage for an agent.

ProjectLocationAgentTestCaseCreateCall
Creates a test case for the given agent.

ProjectLocationAgentTestCaseExportCall
Exports the test cases under the agent to a Cloud Storage bucket or a local file. Filter can be applied to export a subset of test cases.

ProjectLocationAgentTestCaseGetCall
Gets a test case.

ProjectLocationAgentTestCaseImportCall
Imports the test cases from a Cloud Storage bucket or a local file. It always creates new test cases and won't overwite any existing ones. The provided ID in the imported test case is neglected.

ProjectLocationAgentTestCaseListCall
Fetches a list of test cases for a given agent.

ProjectLocationAgentTestCasePatchCall
Updates the specified test case.

ProjectLocationAgentTestCaseResultGetCall
Gets a test case result.

ProjectLocationAgentTestCaseResultListCall
Fetches a list of results for a given test case.

ProjectLocationAgentTestCaseRunCall
Kicks off a test case run.

ProjectLocationAgentValidateCall
Validates the specified agent and creates or updates validation results. The agent in draft version is validated. Please call this API after the training is completed to get the complete validation results.

ProjectLocationAgentWebhookCreateCall
Creates a webhook in the specified agent.

ProjectLocationAgentWebhookDeleteCall
Deletes the specified webhook.

ProjectLocationAgentWebhookGetCall
Retrieves the specified webhook.

ProjectLocationAgentWebhookListCall
Returns the list of all webhooks in the specified agent.

ProjectLocationAgentWebhookPatchCall
Updates the specified webhook.

ProjectLocationOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectLocationSecuritySettingCreateCall
Create security settings in the specified location.

ProjectLocationSecuritySettingDeleteCall
Deletes the specified SecuritySettings.

ProjectLocationSecuritySettingGetCall
Retrieves the specified SecuritySettings. The returned settings may be stale by up to 1 minute.

ProjectLocationSecuritySettingListCall
Returns the list of all security settings in the specified location.

ProjectLocationSecuritySettingPatchCall
Updates the specified SecuritySettings.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Dialogflow hub.

ProjectOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.