Struct google_dialogflow3::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Dialogflow hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_dialogflow3 as dialogflow3;
 
use std::default::Default;
use oauth2;
use dialogflow3::Dialogflow;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Dialogflow::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_agents_create(...)`, `locations_agents_delete(...)`, `locations_agents_entity_types_create(...)`, `locations_agents_entity_types_delete(...)`, `locations_agents_entity_types_get(...)`, `locations_agents_entity_types_list(...)`, `locations_agents_entity_types_patch(...)`, `locations_agents_environments_create(...)`, `locations_agents_environments_delete(...)`, `locations_agents_environments_experiments_create(...)`, `locations_agents_environments_experiments_delete(...)`, `locations_agents_environments_experiments_get(...)`, `locations_agents_environments_experiments_list(...)`, `locations_agents_environments_experiments_patch(...)`, `locations_agents_environments_experiments_start(...)`, `locations_agents_environments_experiments_stop(...)`, `locations_agents_environments_get(...)`, `locations_agents_environments_list(...)`, `locations_agents_environments_lookup_environment_history(...)`, `locations_agents_environments_patch(...)`, `locations_agents_environments_sessions_detect_intent(...)`, `locations_agents_environments_sessions_entity_types_create(...)`, `locations_agents_environments_sessions_entity_types_delete(...)`, `locations_agents_environments_sessions_entity_types_get(...)`, `locations_agents_environments_sessions_entity_types_list(...)`, `locations_agents_environments_sessions_entity_types_patch(...)`, `locations_agents_environments_sessions_fulfill_intent(...)`, `locations_agents_environments_sessions_match_intent(...)`, `locations_agents_export(...)`, `locations_agents_flows_create(...)`, `locations_agents_flows_delete(...)`, `locations_agents_flows_get(...)`, `locations_agents_flows_get_validation_result(...)`, `locations_agents_flows_list(...)`, `locations_agents_flows_pages_create(...)`, `locations_agents_flows_pages_delete(...)`, `locations_agents_flows_pages_get(...)`, `locations_agents_flows_pages_list(...)`, `locations_agents_flows_pages_patch(...)`, `locations_agents_flows_patch(...)`, `locations_agents_flows_train(...)`, `locations_agents_flows_transition_route_groups_create(...)`, `locations_agents_flows_transition_route_groups_delete(...)`, `locations_agents_flows_transition_route_groups_get(...)`, `locations_agents_flows_transition_route_groups_list(...)`, `locations_agents_flows_transition_route_groups_patch(...)`, `locations_agents_flows_validate(...)`, `locations_agents_flows_versions_create(...)`, `locations_agents_flows_versions_delete(...)`, `locations_agents_flows_versions_get(...)`, `locations_agents_flows_versions_list(...)`, `locations_agents_flows_versions_load(...)`, `locations_agents_flows_versions_patch(...)`, `locations_agents_get(...)`, `locations_agents_get_validation_result(...)`, `locations_agents_intents_create(...)`, `locations_agents_intents_delete(...)`, `locations_agents_intents_get(...)`, `locations_agents_intents_list(...)`, `locations_agents_intents_patch(...)`, `locations_agents_list(...)`, `locations_agents_patch(...)`, `locations_agents_restore(...)`, `locations_agents_sessions_detect_intent(...)`, `locations_agents_sessions_entity_types_create(...)`, `locations_agents_sessions_entity_types_delete(...)`, `locations_agents_sessions_entity_types_get(...)`, `locations_agents_sessions_entity_types_list(...)`, `locations_agents_sessions_entity_types_patch(...)`, `locations_agents_sessions_fulfill_intent(...)`, `locations_agents_sessions_match_intent(...)`, `locations_agents_test_cases_batch_delete(...)`, `locations_agents_test_cases_batch_run(...)`, `locations_agents_test_cases_calculate_coverage(...)`, `locations_agents_test_cases_create(...)`, `locations_agents_test_cases_export(...)`, `locations_agents_test_cases_get(...)`, `locations_agents_test_cases_import(...)`, `locations_agents_test_cases_list(...)`, `locations_agents_test_cases_patch(...)`, `locations_agents_test_cases_results_get(...)`, `locations_agents_test_cases_results_list(...)`, `locations_agents_test_cases_run(...)`, `locations_agents_validate(...)`, `locations_agents_webhooks_create(...)`, `locations_agents_webhooks_delete(...)`, `locations_agents_webhooks_get(...)`, `locations_agents_webhooks_list(...)`, `locations_agents_webhooks_patch(...)`, `locations_operations_cancel(...)`, `locations_operations_get(...)`, `locations_operations_list(...)`, `locations_security_settings_create(...)`, `locations_security_settings_delete(...)`, `locations_security_settings_get(...)`, `locations_security_settings_list(...)`, `locations_security_settings_patch(...)`, `operations_cancel(...)`, `operations_get(...)` and `operations_list(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_agents_entity_types_create(
&self,
request: GoogleCloudDialogflowCxV3EntityType,
parent: &str
) -> ProjectLocationAgentEntityTypeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an entity type in the specified agent.

Arguments

request - No description provided.

parent - Required. The agent to create a entity type for. Format: projects//locations//agents/.

pub fn locations_agents_entity_types_delete(
&self,
name: &str
) -> ProjectLocationAgentEntityTypeDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified entity type.

Arguments

name - Required. The name of the entity type to delete. Format: projects//locations//agents//entityTypes/.

pub fn locations_agents_entity_types_get(
&self,
name: &str
) -> ProjectLocationAgentEntityTypeGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified entity type.

Arguments

name - Required. The name of the entity type. Format: projects//locations//agents//entityTypes/.

pub fn locations_agents_entity_types_list(
&self,
parent: &str
) -> ProjectLocationAgentEntityTypeListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all entity types in the specified agent.

Arguments

parent - Required. The agent to list all entity types for. Format: projects//locations//agents/.

pub fn locations_agents_entity_types_patch(
&self,
request: GoogleCloudDialogflowCxV3EntityType,
name: &str
) -> ProjectLocationAgentEntityTypePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified entity type.

Arguments

request - No description provided.

name - The unique identifier of the entity type. Required for EntityTypes.UpdateEntityType. Format: projects//locations//agents//entityTypes/.

pub fn locations_agents_environments_experiments_create(
&self,
request: GoogleCloudDialogflowCxV3Experiment,
parent: &str
) -> ProjectLocationAgentEnvironmentExperimentCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an Experiment in the specified Environment.

Arguments

request - No description provided.

parent - Required. The Agent to create an Environment for. Format: projects//locations//agents//environments/.

pub fn locations_agents_environments_experiments_delete(
&self,
name: &str
) -> ProjectLocationAgentEnvironmentExperimentDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified Experiment.

Arguments

name - Required. The name of the Environment to delete. Format: projects//locations//agents//environments//experiments/.

pub fn locations_agents_environments_experiments_get(
&self,
name: &str
) -> ProjectLocationAgentEnvironmentExperimentGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified Experiment.

Arguments

name - Required. The name of the Environment. Format: projects//locations//agents//environments//experiments/.

pub fn locations_agents_environments_experiments_list(
&self,
parent: &str
) -> ProjectLocationAgentEnvironmentExperimentListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all experiments in the specified Environment.

Arguments

parent - Required. The Environment to list all environments for. Format: projects//locations//agents//environments/.

pub fn locations_agents_environments_experiments_patch(
&self,
request: GoogleCloudDialogflowCxV3Experiment,
name: &str
) -> ProjectLocationAgentEnvironmentExperimentPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified Experiment.

Arguments

request - No description provided.

name - The name of the experiment. Format: projects//locations//agents//environments//experiments/..

pub fn locations_agents_environments_experiments_start(
&self,
request: GoogleCloudDialogflowCxV3StartExperimentRequest,
name: &str
) -> ProjectLocationAgentEnvironmentExperimentStartCall<'a, C>

Create a builder to help you perform the following task:

Starts the specified Experiment. This rpc only changes the state of experiment from PENDING to RUNNING.

Arguments

request - No description provided.

name - Required. Resource name of the experiment to start. Format: projects//locations//agents//environments//experiments/.

pub fn locations_agents_environments_experiments_stop(
&self,
request: GoogleCloudDialogflowCxV3StopExperimentRequest,
name: &str
) -> ProjectLocationAgentEnvironmentExperimentStopCall<'a, C>

Create a builder to help you perform the following task:

Stops the specified Experiment. This rpc only changes the state of experiment from RUNNING to DONE.

Arguments

request - No description provided.

name - Required. Resource name of the experiment to stop. Format: projects//locations//agents//environments//experiments/.

pub fn locations_agents_environments_sessions_entity_types_create(
&self,
request: GoogleCloudDialogflowCxV3SessionEntityType,
parent: &str
) -> ProjectLocationAgentEnvironmentSessionEntityTypeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a session entity type.

Arguments

request - No description provided.

parent - Required. The session to create a session entity type for. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_environments_sessions_entity_types_delete(
&self,
name: &str
) -> ProjectLocationAgentEnvironmentSessionEntityTypeDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified session entity type.

Arguments

name - Required. The name of the session entity type to delete. Format: projects//locations//agents//sessions//entityTypes/ or projects//locations//agents//environments//sessions//entityTypes/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_environments_sessions_entity_types_get(
&self,
name: &str
) -> ProjectLocationAgentEnvironmentSessionEntityTypeGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified session entity type.

Arguments

name - Required. The name of the session entity type. Format: projects//locations//agents//sessions//entityTypes/ or projects//locations//agents//environments//sessions//entityTypes/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_environments_sessions_entity_types_list(
&self,
parent: &str
) -> ProjectLocationAgentEnvironmentSessionEntityTypeListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all session entity types in the specified session.

Arguments

parent - Required. The session to list all session entity types from. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_environments_sessions_entity_types_patch(
&self,
request: GoogleCloudDialogflowCxV3SessionEntityType,
name: &str
) -> ProjectLocationAgentEnvironmentSessionEntityTypePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified session entity type.

Arguments

request - No description provided.

name - Required. The unique identifier of the session entity type. Format: projects//locations//agents//sessions//entityTypes/ or projects//locations//agents//environments//sessions//entityTypes/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_environments_sessions_detect_intent(
&self,
request: GoogleCloudDialogflowCxV3DetectIntentRequest,
session: &str
) -> ProjectLocationAgentEnvironmentSessionDetectIntentCall<'a, C>

Create a builder to help you perform the following task:

Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause session entity types to be updated, which in turn might affect results of future queries. Note: Always use agent versions for production traffic. See Versions and environments.

Arguments

request - No description provided.

session - Required. The name of the session this query is sent to. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment. It's up to the API caller to choose an appropriate Session ID. It can be a random number or some type of session identifiers (preferably hashed). The length of the Session ID must not exceed 36 characters. For more information, see the sessions guide. Note: Always use agent versions for production traffic. See Versions and environments.

pub fn locations_agents_environments_sessions_fulfill_intent(
&self,
request: GoogleCloudDialogflowCxV3FulfillIntentRequest,
session: &str
) -> ProjectLocationAgentEnvironmentSessionFulfillIntentCall<'a, C>

Create a builder to help you perform the following task:

Fulfills a matched intent returned by MatchIntent. Must be called after MatchIntent, with input from MatchIntentResponse. Otherwise, the behavior is undefined.

Arguments

request - No description provided.

session - Required. The name of the session this query is sent to. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment. It's up to the API caller to choose an appropriate Session ID. It can be a random number or some type of session identifiers (preferably hashed). The length of the Session ID must not exceed 36 characters. For more information, see the sessions guide.

pub fn locations_agents_environments_sessions_match_intent(
&self,
request: GoogleCloudDialogflowCxV3MatchIntentRequest,
session: &str
) -> ProjectLocationAgentEnvironmentSessionMatchIntentCall<'a, C>

Create a builder to help you perform the following task:

Returns preliminary intent match results, doesn't change the session status.

Arguments

request - No description provided.

session - Required. The name of the session this query is sent to. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment. It's up to the API caller to choose an appropriate Session ID. It can be a random number or some type of session identifiers (preferably hashed). The length of the Session ID must not exceed 36 characters. For more information, see the sessions guide.

pub fn locations_agents_environments_create(
&self,
request: GoogleCloudDialogflowCxV3Environment,
parent: &str
) -> ProjectLocationAgentEnvironmentCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an Environment in the specified Agent.

Arguments

request - No description provided.

parent - Required. The Agent to create an Environment for. Format: projects//locations//agents/.

pub fn locations_agents_environments_delete(
&self,
name: &str
) -> ProjectLocationAgentEnvironmentDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified Environment.

Arguments

name - Required. The name of the Environment to delete. Format: projects//locations//agents//environments/.

pub fn locations_agents_environments_get(
&self,
name: &str
) -> ProjectLocationAgentEnvironmentGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified Environment.

Arguments

name - Required. The name of the Environment. Format: projects//locations//agents//environments/.

pub fn locations_agents_environments_list(
&self,
parent: &str
) -> ProjectLocationAgentEnvironmentListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all environments in the specified Agent.

Arguments

parent - Required. The Agent to list all environments for. Format: projects//locations//agents/.

pub fn locations_agents_environments_lookup_environment_history(
&self,
name: &str
) -> ProjectLocationAgentEnvironmentLookupEnvironmentHistoryCall<'a, C>

Create a builder to help you perform the following task:

Looks up the history of the specified Environment.

Arguments

name - Required. Resource name of the environment to look up the history for. Format: projects//locations//agents//environments/.

pub fn locations_agents_environments_patch(
&self,
request: GoogleCloudDialogflowCxV3Environment,
name: &str
) -> ProjectLocationAgentEnvironmentPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified Environment.

Arguments

request - No description provided.

name - The name of the environment. Format: projects//locations//agents//environments/.

pub fn locations_agents_flows_pages_create(
&self,
request: GoogleCloudDialogflowCxV3Page,
parent: &str
) -> ProjectLocationAgentFlowPageCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a page in the specified flow.

Arguments

request - No description provided.

parent - Required. The flow to create a page for. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_pages_delete(
&self,
name: &str
) -> ProjectLocationAgentFlowPageDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified page.

Arguments

name - Required. The name of the page to delete. Format: projects//locations//agents//Flows//pages/.

pub fn locations_agents_flows_pages_get(
&self,
name: &str
) -> ProjectLocationAgentFlowPageGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified page.

Arguments

name - Required. The name of the page. Format: projects//locations//agents//flows//pages/.

pub fn locations_agents_flows_pages_list(
&self,
parent: &str
) -> ProjectLocationAgentFlowPageListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all pages in the specified flow.

Arguments

parent - Required. The flow to list all pages for. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_pages_patch(
&self,
request: GoogleCloudDialogflowCxV3Page,
name: &str
) -> ProjectLocationAgentFlowPagePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified page.

Arguments

request - No description provided.

name - The unique identifier of the page. Required for the Pages.UpdatePage method. Pages.CreatePage populates the name automatically. Format: projects//locations//agents//flows//pages/.

pub fn locations_agents_flows_transition_route_groups_create(
&self,
request: GoogleCloudDialogflowCxV3TransitionRouteGroup,
parent: &str
) -> ProjectLocationAgentFlowTransitionRouteGroupCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an TransitionRouteGroup in the specified flow.

Arguments

request - No description provided.

parent - Required. The flow to create an TransitionRouteGroup for. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_transition_route_groups_delete(
&self,
name: &str
) -> ProjectLocationAgentFlowTransitionRouteGroupDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified TransitionRouteGroup.

Arguments

name - Required. The name of the TransitionRouteGroup to delete. Format: projects//locations//agents//flows//transitionRouteGroups/.

pub fn locations_agents_flows_transition_route_groups_get(
&self,
name: &str
) -> ProjectLocationAgentFlowTransitionRouteGroupGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified TransitionRouteGroup.

Arguments

name - Required. The name of the TransitionRouteGroup. Format: projects//locations//agents//flows//transitionRouteGroups/.

pub fn locations_agents_flows_transition_route_groups_list(
&self,
parent: &str
) -> ProjectLocationAgentFlowTransitionRouteGroupListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all transition route groups in the specified flow.

Arguments

parent - Required. The flow to list all transition route groups for. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_transition_route_groups_patch(
&self,
request: GoogleCloudDialogflowCxV3TransitionRouteGroup,
name: &str
) -> ProjectLocationAgentFlowTransitionRouteGroupPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified TransitionRouteGroup.

Arguments

request - No description provided.

name - The unique identifier of the transition route group. TransitionRouteGroups.CreateTransitionRouteGroup populates the name automatically. Format: projects//locations//agents//flows//transitionRouteGroups/.

pub fn locations_agents_flows_versions_create(
&self,
request: GoogleCloudDialogflowCxV3Version,
parent: &str
) -> ProjectLocationAgentFlowVersionCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a Version in the specified Flow.

Arguments

request - No description provided.

parent - Required. The Flow to create an Version for. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_versions_delete(
&self,
name: &str
) -> ProjectLocationAgentFlowVersionDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified Version.

Arguments

name - Required. The name of the Version to delete. Format: projects//locations//agents//flows//versions/.

pub fn locations_agents_flows_versions_get(
&self,
name: &str
) -> ProjectLocationAgentFlowVersionGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified Version.

Arguments

name - Required. The name of the Version. Format: projects//locations//agents//flows//versions/.

pub fn locations_agents_flows_versions_list(
&self,
parent: &str
) -> ProjectLocationAgentFlowVersionListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all versions in the specified Flow.

Arguments

parent - Required. The Flow to list all versions for. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_versions_load(
&self,
request: GoogleCloudDialogflowCxV3LoadVersionRequest,
name: &str
) -> ProjectLocationAgentFlowVersionLoadCall<'a, C>

Create a builder to help you perform the following task:

Loads a specified version to draft version.

Arguments

request - No description provided.

name - Required. The Version to be loaded to draft version. Format: projects//locations//agents//flows//versions/.

pub fn locations_agents_flows_versions_patch(
&self,
request: GoogleCloudDialogflowCxV3Version,
name: &str
) -> ProjectLocationAgentFlowVersionPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified Version.

Arguments

request - No description provided.

name - Format: projects//locations//agents//flows//versions/. Version ID is a self-increasing number generated by Dialogflow upon version creation.

pub fn locations_agents_flows_create(
&self,
request: GoogleCloudDialogflowCxV3Flow,
parent: &str
) -> ProjectLocationAgentFlowCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a flow in the specified agent.

Arguments

request - No description provided.

parent - Required. The agent to create a flow for. Format: projects//locations//agents/.

pub fn locations_agents_flows_delete(
&self,
name: &str
) -> ProjectLocationAgentFlowDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a specified flow.

Arguments

name - Required. The name of the flow to delete. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_get(
&self,
name: &str
) -> ProjectLocationAgentFlowGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified flow.

Arguments

name - Required. The name of the flow to get. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_get_validation_result(
&self,
name: &str
) -> ProjectLocationAgentFlowGetValidationResultCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest flow validation result. Flow validation is performed when ValidateFlow is called.

Arguments

name - Required. The flow name. Format: projects//locations//agents//flows//validationResult.

pub fn locations_agents_flows_list(
&self,
parent: &str
) -> ProjectLocationAgentFlowListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all flows in the specified agent.

Arguments

parent - Required. The agent containing the flows. Format: projects//locations//agents/.

pub fn locations_agents_flows_patch(
&self,
request: GoogleCloudDialogflowCxV3Flow,
name: &str
) -> ProjectLocationAgentFlowPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified flow.

Arguments

request - No description provided.

name - The unique identifier of the flow. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_train(
&self,
request: GoogleCloudDialogflowCxV3TrainFlowRequest,
name: &str
) -> ProjectLocationAgentFlowTrainCall<'a, C>

Create a builder to help you perform the following task:

Trains the specified flow. Note that only the flow in 'draft' environment is trained.

Arguments

request - No description provided.

name - Required. The flow to train. Format: projects//locations//agents//flows/.

pub fn locations_agents_flows_validate(
&self,
request: GoogleCloudDialogflowCxV3ValidateFlowRequest,
name: &str
) -> ProjectLocationAgentFlowValidateCall<'a, C>

Create a builder to help you perform the following task:

Validates the specified flow and creates or updates validation results. Please call this API after the training is completed to get the complete validation results.

Arguments

request - No description provided.

name - Required. The flow to validate. Format: projects//locations//agents//flows/.

pub fn locations_agents_intents_create(
&self,
request: GoogleCloudDialogflowCxV3Intent,
parent: &str
) -> ProjectLocationAgentIntentCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an intent in the specified agent.

Arguments

request - No description provided.

parent - Required. The agent to create an intent for. Format: projects//locations//agents/.

pub fn locations_agents_intents_delete(
&self,
name: &str
) -> ProjectLocationAgentIntentDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified intent.

Arguments

name - Required. The name of the intent to delete. Format: projects//locations//agents//intents/.

pub fn locations_agents_intents_get(
&self,
name: &str
) -> ProjectLocationAgentIntentGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified intent.

Arguments

name - Required. The name of the intent. Format: projects//locations//agents//intents/.

pub fn locations_agents_intents_list(
&self,
parent: &str
) -> ProjectLocationAgentIntentListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all intents in the specified agent.

Arguments

parent - Required. The agent to list all intents for. Format: projects//locations//agents/.

pub fn locations_agents_intents_patch(
&self,
request: GoogleCloudDialogflowCxV3Intent,
name: &str
) -> ProjectLocationAgentIntentPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified intent.

Arguments

request - No description provided.

name - The unique identifier of the intent. Required for the Intents.UpdateIntent method. Intents.CreateIntent populates the name automatically. Format: projects//locations//agents//intents/.

pub fn locations_agents_sessions_entity_types_create(
&self,
request: GoogleCloudDialogflowCxV3SessionEntityType,
parent: &str
) -> ProjectLocationAgentSessionEntityTypeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a session entity type.

Arguments

request - No description provided.

parent - Required. The session to create a session entity type for. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_sessions_entity_types_delete(
&self,
name: &str
) -> ProjectLocationAgentSessionEntityTypeDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified session entity type.

Arguments

name - Required. The name of the session entity type to delete. Format: projects//locations//agents//sessions//entityTypes/ or projects//locations//agents//environments//sessions//entityTypes/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_sessions_entity_types_get(
&self,
name: &str
) -> ProjectLocationAgentSessionEntityTypeGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified session entity type.

Arguments

name - Required. The name of the session entity type. Format: projects//locations//agents//sessions//entityTypes/ or projects//locations//agents//environments//sessions//entityTypes/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_sessions_entity_types_list(
&self,
parent: &str
) -> ProjectLocationAgentSessionEntityTypeListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all session entity types in the specified session.

Arguments

parent - Required. The session to list all session entity types from. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_sessions_entity_types_patch(
&self,
request: GoogleCloudDialogflowCxV3SessionEntityType,
name: &str
) -> ProjectLocationAgentSessionEntityTypePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified session entity type.

Arguments

request - No description provided.

name - Required. The unique identifier of the session entity type. Format: projects//locations//agents//sessions//entityTypes/ or projects//locations//agents//environments//sessions//entityTypes/. If Environment ID is not specified, we assume default 'draft' environment.

pub fn locations_agents_sessions_detect_intent(
&self,
request: GoogleCloudDialogflowCxV3DetectIntentRequest,
session: &str
) -> ProjectLocationAgentSessionDetectIntentCall<'a, C>

Create a builder to help you perform the following task:

Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause session entity types to be updated, which in turn might affect results of future queries. Note: Always use agent versions for production traffic. See Versions and environments.

Arguments

request - No description provided.

session - Required. The name of the session this query is sent to. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment. It's up to the API caller to choose an appropriate Session ID. It can be a random number or some type of session identifiers (preferably hashed). The length of the Session ID must not exceed 36 characters. For more information, see the sessions guide. Note: Always use agent versions for production traffic. See Versions and environments.

pub fn locations_agents_sessions_fulfill_intent(
&self,
request: GoogleCloudDialogflowCxV3FulfillIntentRequest,
session: &str
) -> ProjectLocationAgentSessionFulfillIntentCall<'a, C>

Create a builder to help you perform the following task:

Fulfills a matched intent returned by MatchIntent. Must be called after MatchIntent, with input from MatchIntentResponse. Otherwise, the behavior is undefined.

Arguments

request - No description provided.

session - Required. The name of the session this query is sent to. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment. It's up to the API caller to choose an appropriate Session ID. It can be a random number or some type of session identifiers (preferably hashed). The length of the Session ID must not exceed 36 characters. For more information, see the sessions guide.

pub fn locations_agents_sessions_match_intent(
&self,
request: GoogleCloudDialogflowCxV3MatchIntentRequest,
session: &str
) -> ProjectLocationAgentSessionMatchIntentCall<'a, C>

Create a builder to help you perform the following task:

Returns preliminary intent match results, doesn't change the session status.

Arguments

request - No description provided.

session - Required. The name of the session this query is sent to. Format: projects//locations//agents//sessions/ or projects//locations//agents//environments//sessions/. If Environment ID is not specified, we assume default 'draft' environment. It's up to the API caller to choose an appropriate Session ID. It can be a random number or some type of session identifiers (preferably hashed). The length of the Session ID must not exceed 36 characters. For more information, see the sessions guide.

pub fn locations_agents_test_cases_results_get(
&self,
name: &str
) -> ProjectLocationAgentTestCaseResultGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a test case result.

Arguments

name - Required. The name of the testcase. Format: projects//locations//agents//testCases//results/.

pub fn locations_agents_test_cases_results_list(
&self,
parent: &str
) -> ProjectLocationAgentTestCaseResultListCall<'a, C>

Create a builder to help you perform the following task:

Fetches a list of results for a given test case.

Arguments

parent - Required. The test case to list results for. Format: projects//locations//agents// testCases/. Specify a - as a wildcard for TestCase ID to list results across multiple test cases.

pub fn locations_agents_test_cases_batch_delete(
&self,
request: GoogleCloudDialogflowCxV3BatchDeleteTestCasesRequest,
parent: &str
) -> ProjectLocationAgentTestCaseBatchDeleteCall<'a, C>

Create a builder to help you perform the following task:

Batch deletes test cases.

Arguments

request - No description provided.

parent - Required. The agent to delete test cases from. Format: projects//locations//agents/.

pub fn locations_agents_test_cases_batch_run(
&self,
request: GoogleCloudDialogflowCxV3BatchRunTestCasesRequest,
parent: &str
) -> ProjectLocationAgentTestCaseBatchRunCall<'a, C>

Create a builder to help you perform the following task:

Kicks off a batch run of test cases.

Arguments

request - No description provided.

parent - Required. Agent name. Format: projects//locations//agents/ .

pub fn locations_agents_test_cases_calculate_coverage(
&self,
agent: &str
) -> ProjectLocationAgentTestCaseCalculateCoverageCall<'a, C>

Create a builder to help you perform the following task:

Calculates the test coverage for an agent.

Arguments

agent - Required. The agent to calculate coverage for. Format: projects//locations//agents/.

pub fn locations_agents_test_cases_create(
&self,
request: GoogleCloudDialogflowCxV3TestCase,
parent: &str
) -> ProjectLocationAgentTestCaseCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a test case for the given agent.

Arguments

request - No description provided.

parent - Required. The agent to create the test case for. Format: projects//locations//agents/.

pub fn locations_agents_test_cases_export(
&self,
request: GoogleCloudDialogflowCxV3ExportTestCasesRequest,
parent: &str
) -> ProjectLocationAgentTestCaseExportCall<'a, C>

Create a builder to help you perform the following task:

Exports the test cases under the agent to a Cloud Storage bucket or a local file. Filter can be applied to export a subset of test cases.

Arguments

request - No description provided.

parent - Required. The agent where to export test cases from. Format: projects//locations//agents/.

pub fn locations_agents_test_cases_get(
&self,
name: &str
) -> ProjectLocationAgentTestCaseGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a test case.

Arguments

name - Required. The name of the testcase. Format: projects//locations//agents//testCases/.

pub fn locations_agents_test_cases_import(
&self,
request: GoogleCloudDialogflowCxV3ImportTestCasesRequest,
parent: &str
) -> ProjectLocationAgentTestCaseImportCall<'a, C>

Create a builder to help you perform the following task:

Imports the test cases from a Cloud Storage bucket or a local file. It always creates new test cases and won't overwite any existing ones. The provided ID in the imported test case is neglected.

Arguments

request - No description provided.

parent - Required. The agent to import test cases to. Format: projects//locations//agents/.

pub fn locations_agents_test_cases_list(
&self,
parent: &str
) -> ProjectLocationAgentTestCaseListCall<'a, C>

Create a builder to help you perform the following task:

Fetches a list of test cases for a given agent.

Arguments

parent - Required. The agent to list all pages for. Format: projects//locations//agents/.

pub fn locations_agents_test_cases_patch(
&self,
request: GoogleCloudDialogflowCxV3TestCase,
name: &str
) -> ProjectLocationAgentTestCasePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified test case.

Arguments

request - No description provided.

name - The unique identifier of the test case. TestCases.CreateTestCase will populate the name automatically. Otherwise use format: projects//locations//agents/ /testCases/.

pub fn locations_agents_test_cases_run(
&self,
request: GoogleCloudDialogflowCxV3RunTestCaseRequest,
name: &str
) -> ProjectLocationAgentTestCaseRunCall<'a, C>

Create a builder to help you perform the following task:

Kicks off a test case run.

Arguments

request - No description provided.

name - Required. Format of test case name to run: projects//locations/ /agents//testCases/.

pub fn locations_agents_webhooks_create(
&self,
request: GoogleCloudDialogflowCxV3Webhook,
parent: &str
) -> ProjectLocationAgentWebhookCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a webhook in the specified agent.

Arguments

request - No description provided.

parent - Required. The agent to create a webhook for. Format: projects//locations//agents/.

pub fn locations_agents_webhooks_delete(
&self,
name: &str
) -> ProjectLocationAgentWebhookDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified webhook.

Arguments

name - Required. The name of the webhook to delete. Format: projects//locations//agents//webhooks/.

pub fn locations_agents_webhooks_get(
&self,
name: &str
) -> ProjectLocationAgentWebhookGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified webhook.

Arguments

name - Required. The name of the webhook. Format: projects//locations//agents//webhooks/.

pub fn locations_agents_webhooks_list(
&self,
parent: &str
) -> ProjectLocationAgentWebhookListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all webhooks in the specified agent.

Arguments

parent - Required. The agent to list all webhooks for. Format: projects//locations//agents/.

pub fn locations_agents_webhooks_patch(
&self,
request: GoogleCloudDialogflowCxV3Webhook,
name: &str
) -> ProjectLocationAgentWebhookPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified webhook.

Arguments

request - No description provided.

name - The unique identifier of the webhook. Required for the Webhooks.UpdateWebhook method. Webhooks.CreateWebhook populates the name automatically. Format: projects//locations//agents//webhooks/.

pub fn locations_agents_create(
&self,
request: GoogleCloudDialogflowCxV3Agent,
parent: &str
) -> ProjectLocationAgentCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an agent in the specified location.

Arguments

request - No description provided.

parent - Required. The location to create a agent for. Format: projects//locations/.

pub fn locations_agents_delete(
&self,
name: &str
) -> ProjectLocationAgentDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified agent.

Arguments

name - Required. The name of the agent to delete. Format: projects//locations//agents/.

pub fn locations_agents_export(
&self,
request: GoogleCloudDialogflowCxV3ExportAgentRequest,
name: &str
) -> ProjectLocationAgentExportCall<'a, C>

Create a builder to help you perform the following task:

Exports the specified agent to a binary file.

Arguments

request - No description provided.

name - Required. The name of the agent to export. Format: projects//locations//agents/.

pub fn locations_agents_get(
&self,
name: &str
) -> ProjectLocationAgentGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified agent.

Arguments

name - Required. The name of the agent. Format: projects//locations//agents/.

pub fn locations_agents_get_validation_result(
&self,
name: &str
) -> ProjectLocationAgentGetValidationResultCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest agent validation result. Agent validation is performed when ValidateAgent is called.

Arguments

name - Required. The agent name. Format: projects//locations//agents//validationResult.

pub fn locations_agents_list(
&self,
parent: &str
) -> ProjectLocationAgentListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all agents in the specified location.

Arguments

parent - Required. The location to list all agents for. Format: projects//locations/.

pub fn locations_agents_patch(
&self,
request: GoogleCloudDialogflowCxV3Agent,
name: &str
) -> ProjectLocationAgentPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified agent.

Arguments

request - No description provided.

name - The unique identifier of the agent. Required for the Agents.UpdateAgent method. Agents.CreateAgent populates the name automatically. Format: projects//locations//agents/.

pub fn locations_agents_restore(
&self,
request: GoogleCloudDialogflowCxV3RestoreAgentRequest,
name: &str
) -> ProjectLocationAgentRestoreCall<'a, C>

Create a builder to help you perform the following task:

Restores the specified agent from a binary file. Replaces the current agent with a new one. Note that all existing resources in agent (e.g. intents, entity types, flows) will be removed.

Arguments

request - No description provided.

name - Required. The name of the agent to restore into. Format: projects//locations//agents/.

pub fn locations_agents_validate(
&self,
request: GoogleCloudDialogflowCxV3ValidateAgentRequest,
name: &str
) -> ProjectLocationAgentValidateCall<'a, C>

Create a builder to help you perform the following task:

Validates the specified agent and creates or updates validation results. The agent in draft version is validated. Please call this API after the training is completed to get the complete validation results.

Arguments

request - No description provided.

name - Required. The agent to validate. Format: projects//locations//agents/.

pub fn locations_operations_cancel(
&self,
name: &str
) -> ProjectLocationOperationCancelCall<'a, C>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

name - The name of the operation resource to be cancelled.

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_operations_list(
&self,
name: &str
) -> ProjectLocationOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_security_settings_create(
&self,
request: GoogleCloudDialogflowCxV3SecuritySettings,
parent: &str
) -> ProjectLocationSecuritySettingCreateCall<'a, C>

Create a builder to help you perform the following task:

Create security settings in the specified location.

Arguments

request - No description provided.

parent - Required. The location to create an SecuritySettings for. Format: projects//locations/.

pub fn locations_security_settings_delete(
&self,
name: &str
) -> ProjectLocationSecuritySettingDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified SecuritySettings.

Arguments

name - Required. The name of the SecuritySettings to delete. Format: projects//locations//securitySettings/.

pub fn locations_security_settings_get(
&self,
name: &str
) -> ProjectLocationSecuritySettingGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified SecuritySettings. The returned settings may be stale by up to 1 minute.

Arguments

name - Required. Resource name of the settings. Format: projects//locations//securitySettings/.

pub fn locations_security_settings_list(
&self,
parent: &str
) -> ProjectLocationSecuritySettingListCall<'a, C>

Create a builder to help you perform the following task:

Returns the list of all security settings in the specified location.

Arguments

parent - Required. The location to list all security settings for. Format: projects//locations/.

pub fn locations_security_settings_patch(
&self,
request: GoogleCloudDialogflowCxV3SecuritySettings,
name: &str
) -> ProjectLocationSecuritySettingPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified SecuritySettings.

Arguments

request - No description provided.

name - Required. Resource name of the settings. Format: projects//locations//securitySettings/.

pub fn operations_cancel(&self, name: &str) -> ProjectOperationCancelCall<'a, C>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

name - The name of the operation resource to be cancelled.

pub fn operations_get(&self, name: &str) -> ProjectOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn operations_list(&self, name: &str) -> ProjectOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.